refactor: drop dead imports and move local-image-model validation out of the route
